refactor(tests): consolidate update tests into single update_test.rs
Merges update_status_test.rs and update_check_test.rs into one file.
Both cover the same module and share tempfile/anyhow infrastructure.
Also fixes the pre-release test to exercise the interval-elapsed path
(adds write_last_checked call) rather than the first-run path.
